requirements.txt
numpy
scipy
astropy
matplotlib
scikit-image
imageio
photutils
pillow
pyraf